alphachatgpt/examples/data/image/requirements.txt
# pdf2image
Pillow
# PyPDF2==2.12.1
pytesseract
# PyMuPDF
# opencv-python